fixed folder structure for commands
