+++ Team-3749-2025/src/main/java/frc/robot/commands/arm/algae/AlgaeMoveUp.java
@@ -1,4 +1,4 @@
-package frc.robot.commands.arm.algaeArmCommands;
+package frc.robot.commands.arm.algae;
 
 import edu.wpi.first.wpilibj2.command.Command;
 import frc.robot.Robot;
