+++ Team-3749-2025/src/main/java/frc/robot/commands/arm/climb/Climb.java
@@ -1,4 +1,4 @@
-package frc.robot.commands.arm.climbArmCommands;
+package frc.robot.commands.arm.climb;
 
 import edu.wpi.first.wpilibj2.command.Command;
 import frc.robot.Robot;
@@ -6,18 +6,18 @@ import frc.robot.subsystems.arm.ArmConstants.climbArmConstants;
 import frc.robot.subsystems.arm.ArmConstants.climbArmConstants.ArmStates;
 import frc.robot.utils.UtilityFunctions;
 
-public class ClimbArmPrepareForClimb extends Command {
+public class Climb extends Command {
 
-    public ClimbArmPrepareForClimb() {
-        super.addRequirements(Robot.climbArm);
+    public Climb() {
+        super.addRequirements(Robot.algaeArm);
     }
 
     @Override
     public void initialize() {
-        Robot.climbArm.setState(ArmStates.PREPARE_FOR_CLIMB);
+        Robot.climbArm.setState(ArmStates.CLIMB);
     }
 
-    @Override
+        @Override
     public void end(boolean interupted) {
     }
