+++ Team-3749-2025/src/main/java/frc/robot/commands/arm/climb/ClimbArmPrepareForClimb.java
@@ -1,4 +1,4 @@
-package frc.robot.commands.arm.climbArmCommands;
+package frc.robot.commands.arm.climb;
 
 import edu.wpi.first.wpilibj2.command.Command;
 import frc.robot.Robot;
@@ -6,24 +6,24 @@ import frc.robot.subsystems.arm.ArmConstants.climbArmConstants;
 import frc.robot.subsystems.arm.ArmConstants.climbArmConstants.ArmStates;
 import frc.robot.utils.UtilityFunctions;
 
-public class ClimbStow extends Command {
+public class ClimbArmPrepareForClimb extends Command {
 
-    public ClimbStow() {
+    public ClimbArmPrepareForClimb() {
         super.addRequirements(Robot.climbArm);
     }
 
     @Override
     public void initialize() {
-        Robot.climbArm.setState(ArmStates.STOWED);
+        Robot.climbArm.setState(ArmStates.PREPARE_FOR_CLIMB);
     }
 
-        @Override
+    @Override
     public void end(boolean interupted) {
     }
 
     @Override
     public boolean isFinished() {
         double position = Robot.climbArm.getPositionRad();
-        return UtilityFunctions.withinMargin(0.001, climbArmConstants.stowSetPoint_rad, position);
+        return UtilityFunctions.withinMargin(0.001, climbArmConstants.climbSetPoint_rad, position);
     }
 }
