+++ Team-3749-2025/src/main/java/frc/robot/commands/arm/climb/ClimbMoveUp.java
@@ -1,18 +1,18 @@
-package frc.robot.commands.arm.climbArmCommands;
+package frc.robot.commands.arm.climb;
 
 import edu.wpi.first.wpilibj2.command.Command;
 import frc.robot.Robot;
 import frc.robot.subsystems.arm.ArmConstants.climbArmConstants.ArmStates;
 
-public class ClimbMoveDown extends Command {
+public class ClimbMoveUp extends Command {
 
-    public ClimbMoveDown() {
+    public ClimbMoveUp() {
         super.addRequirements(Robot.climbArm);
     }
 
     @Override
     public void initialize() {
-        Robot.climbArm.setState(ArmStates.MOVING_DOWN);
+        Robot.climbArm.setState(ArmStates.MOVING_UP);
     }
 
         @Override
