+++ Team-3749-2025/src/main/java/frc/robot/commands/arm/coral/CoralHandOff.java
@@ -1,4 +1,4 @@
-package frc.robot.commands.arm.coralArmCommands;
+package frc.robot.commands.arm.coral;
 
 import edu.wpi.first.wpilibj2.command.Command;
 import frc.robot.Robot;
@@ -6,24 +6,24 @@ import frc.robot.subsystems.arm.ArmConstants.coralArmConstants;
 import frc.robot.subsystems.arm.ArmConstants.coralArmConstants.ArmStates;
 import frc.robot.utils.UtilityFunctions;
 
-public class CoralStow extends Command {
+public class CoralHandOff extends Command {
 
-    public CoralStow() {
+    public CoralHandOff() {
         super.addRequirements(Robot.coralArm);
     }
 
     @Override
     public void initialize() {
-        Robot.coralArm.setState(ArmStates.STOWED);
+        Robot.coralArm.setState(ArmStates.HAND_OFF);
     }
 
-        @Override
+    @Override
     public void end(boolean interupted) {
     }
 
     @Override
     public boolean isFinished() {
         double position = Robot.coralArm.getPositionRad();
-        return UtilityFunctions.withinMargin(0.001, coralArmConstants.stowSetPoint_rad, position);
+        return UtilityFunctions.withinMargin(0.001, coralArmConstants.handOffSetPoint_rad, position);
     }
 }
