+++ Team-3749-2025/src/main/java/frc/robot/commands/arm/coral/CoralPickup.java
@@ -1,4 +1,4 @@
-package frc.robot.commands.arm.coralArmCommands;
+package frc.robot.commands.arm.coral;
 
 import edu.wpi.first.wpilibj2.command.Command;
 import frc.robot.Robot;
@@ -6,24 +6,24 @@ import frc.robot.subsystems.arm.ArmConstants.coralArmConstants;
 import frc.robot.subsystems.arm.ArmConstants.coralArmConstants.ArmStates;
 import frc.robot.utils.UtilityFunctions;
 
-public class CoralHandOff extends Command {
+public class CoralPickup extends Command {
 
-    public CoralHandOff() {
+    public CoralPickup() {
         super.addRequirements(Robot.coralArm);
     }
 
     @Override
     public void initialize() {
-        Robot.coralArm.setState(ArmStates.HAND_OFF);
+        Robot.coralArm.setState(ArmStates.CORAL_PICKUP);
     }
 
-    @Override
+        @Override
     public void end(boolean interupted) {
     }
 
     @Override
     public boolean isFinished() {
         double position = Robot.coralArm.getPositionRad();
-        return UtilityFunctions.withinMargin(0.001, coralArmConstants.handOffSetPoint_rad, position);
+        return UtilityFunctions.withinMargin(0.001, coralArmConstants.coralPickUpSetPoint_Rad, position);
     }
 }
