+++ Team-3749-2025/src/main/java/frc/robot/commands/arm/coral/CoralStow.java
@@ -1,4 +1,4 @@
-package frc.robot.commands.arm.coralArmCommands;
+package frc.robot.commands.arm.coral;
 
 import edu.wpi.first.wpilibj2.command.Command;
 import frc.robot.Robot;
@@ -6,15 +6,15 @@ import frc.robot.subsystems.arm.ArmConstants.coralArmConstants;
 import frc.robot.subsystems.arm.ArmConstants.coralArmConstants.ArmStates;
 import frc.robot.utils.UtilityFunctions;
 
-public class CoralPickup extends Command {
+public class CoralStow extends Command {
 
-    public CoralPickup() {
+    public CoralStow() {
         super.addRequirements(Robot.coralArm);
     }
 
     @Override
     public void initialize() {
-        Robot.coralArm.setState(ArmStates.CORAL_PICKUP);
+        Robot.coralArm.setState(ArmStates.STOWED);
     }
 
         @Override
@@ -24,6 +24,6 @@ public class CoralPickup extends Command {
     @Override
     public boolean isFinished() {
         double position = Robot.coralArm.getPositionRad();
-        return UtilityFunctions.withinMargin(0.001, coralArmConstants.coralPickUpSetPoint_Rad, position);
+        return UtilityFunctions.withinMargin(0.001, coralArmConstants.stowSetPoint_rad, position);
     }
 }
