+++ Team-3749-2025/src/main/java/frc/robot/utils/JoystickIO.java
@@ -5,10 +5,10 @@ import edu.wpi.first.wpilibj2.command.Command;
 import edu.wpi.first.wpilibj2.command.Commands;
 import edu.wpi.first.wpilibj2.command.button.CommandXboxController;
 import frc.robot.Robot;
-import frc.robot.commands.arm.coralArmCommands.CoralMoveDown;
-import frc.robot.commands.arm.coralArmCommands.CoralMoveUp;
-import frc.robot.commands.arm.coralArmCommands.CoralPickup;
-import frc.robot.commands.arm.coralArmCommands.CoralStow;
+import frc.robot.commands.arm.coral.CoralMoveDown;
+import frc.robot.commands.arm.coral.CoralMoveUp;
+import frc.robot.commands.arm.coral.CoralPickup;
+import frc.robot.commands.arm.coral.CoralStow;
 import frc.robot.commands.example.ExampleSubsystemCommand;
 import frc.robot.commands.swerve.DriveStraight;
 import frc.robot.commands.swerve.SwerveDefaultCommand;
